refactor: move recalculate score to root utils package
- condensed recalculate score tests
- pruned some util functions that were only used 1 time and more than
likely won't be used anywhere else in the near future
